# tp4-ecommerce/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ecommerce API",
        "description": "API for Ecommerce application",
        "version": "1.0.0"
    },
    "paths": {
...
        "/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get user's orders",
                "operationId": "325fd3d55bbe9bb39645269928fa4983",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Order"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create a new order",
                "operationId": "d690d6834318e9b064df441e854de8b9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "customer_first_name",
                                    "customer_last_name",
                                    "customer_email",
                                    "customer_phone",
                                    "shipping_address",
                                    "shipping_city",
                                    "shipping_zipcode",
                                    "shipping_country",
                                    "payment_method"
                                ],
                                "properties": {
                                    "customer_first_name": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "customer_last_name": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "customer_email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "customer_phone": {
                                        "type": "string",
                                        "example": "+1234567890"
                                    },
                                    "shipping_address": {
                                        "type": "string",
                                        "example": "123 Main St"
                                    },
                                    "shipping_city": {
                                        "type": "string",
                                        "example": "New York"
                                    },
                                    "shipping_zipcode": {
                                        "type": "string",
                                        "example": "10001"
                                    },
                                    "shipping_country": {
                                        "type": "string",
                                        "example": "USA"
                                    },
                                    "billing_address": {
                                        "type": "string",
                                        "example": "123 Main St"
                                    },
                                    "billing_city": {
                                        "type": "string",
                                        "example": "New York"
                                    },
                                    "billing_zipcode": {
                                        "type": "string",
                                        "example": "10001"
                                    },
                                    "billing_country": {
                                        "type": "string",
                                        "example": "USA"
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "example": "credit_card"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Please deliver in the morning"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "order": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get order details",
                "operationId": "c29621029e70f65247909d3e09031118",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Order": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "order_number": {
                        "type": "string",
                        "example": "ORD-20231215-ABC123"
                    },
                    "status": {
                        "type": "string",
                        "example": "pending"
                    },
                    "subtotal": {
                        "type": "number",
                        "format": "float",
                        "example": 199.98
                    },
                    "shipping": {
                        "type": "number",
                        "format": "float",
                        "example": 10
                    },
                    "tax": {
                        "type": "number",
                        "format": "float",
                        "example": 20
                    },
                    "total": {
                        "type": "number",
                        "format": "float",
                        "example": 229.98
                    },
                    "customer_email": {
                        "type": "string",
                        "example": "john@example.com"
                    },
                    "customer_first_name": {
                        "type": "string",
                        "example": "John"
                    },
                    "customer_last_name": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "shipping_address": {
                        "type": "string",
                        "example": "123 Main St"
                    },
                    "shipping_city": {
                        "type": "string",
                        "example": "New York"
                    },
                    "shipping_zipcode": {
                        "type": "string",
                        "example": "10001"
                    },
                    "shipping_country": {
                        "type": "string",
                        "example": "USA"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}